refactor to put instantiation of builders into respective modules
