+++ great_expectations/rule_based_profiler/expectation_configuration_builder/__init__.py
@@ -1,7 +1,7 @@
 from great_expectations.rule_based_profiler.expectation_configuration_builder.expectation_configuration_builder import (  # isort:skip
     ExpectationConfigurationBuilder,
+    init_rule_expectation_configuration_builders,
 )
-
 from great_expectations.rule_based_profiler.expectation_configuration_builder.default_expectation_configuration_builder import (  # isort:skip
     DefaultExpectationConfigurationBuilder,
 )
