+++ great_expectations/rule_based_profiler/expectation_configuration_builder/expectation_configuration_builder.py
@@ -4,12 +4,15 @@ from typing import Dict, List, Optional, Set, Union
 
 from great_expectations.core.batch import Batch, BatchRequest, RuntimeBatchRequest
 from great_expectations.core.expectation_configuration import ExpectationConfiguration
+from great_expectations.data_context.util import instantiate_class_from_config
 from great_expectations.rule_based_profiler.config import ParameterBuilderConfig
 from great_expectations.rule_based_profiler.helpers.util import (
-    init_rule_parameter_builders,
     set_batch_list_or_batch_request_on_builder,
 )
-from great_expectations.rule_based_profiler.parameter_builder import ParameterBuilder
+from great_expectations.rule_based_profiler.parameter_builder import (
+    ParameterBuilder,
+    init_rule_parameter_builders,
+)
 from great_expectations.rule_based_profiler.types import (
     Builder,
     Domain,
@@ -139,3 +142,39 @@ class ExpectationConfigurationBuilder(Builder, ABC):
     @property
     def validation_parameter_builders(self) -> Optional[List[ParameterBuilder]]:
         return self._validation_parameter_builders
+
+
+def init_rule_expectation_configuration_builders(
+    expectation_configuration_builder_configs: List[dict],
+    data_context: Optional["DataContext"] = None,  # noqa: F821
+) -> List["ExpectationConfigurationBuilder"]:  # noqa: F821
+    expectation_configuration_builder_config: dict
+    return [
+        init_expectation_configuration_builder(
+            expectation_configuration_builder_config=expectation_configuration_builder_config,
+            data_context=data_context,
+        )
+        for expectation_configuration_builder_config in expectation_configuration_builder_configs
+    ]
+
+
+def init_expectation_configuration_builder(
+    expectation_configuration_builder_config: Union[
+        "ExpectationConfigurationBuilder", dict  # noqa: F821
+    ],
+    data_context: Optional["DataContext"] = None,  # noqa: F821
+) -> "ExpectationConfigurationBuilder":  # noqa: F821
+    if not isinstance(expectation_configuration_builder_config, dict):
+        expectation_configuration_builder_config = (
+            expectation_configuration_builder_config.to_dict()
+        )
+
+    expectation_configuration_builder: "ExpectationConfigurationBuilder" = instantiate_class_from_config(  # noqa: F821
+        config=expectation_configuration_builder_config,
+        runtime_environment={"data_context": data_context},
+        config_defaults={
+            "class_name": "DefaultExpectationConfigurationBuilder",
+            "module_name": "great_expectations.rule_based_profiler.expectation_configuration_builder",
+        },
+    )
+    return expectation_configuration_builder
